requirements.txt
pillow>=8.3.2
django-widget-tweaks>=1.4.8